+++ background.tsx
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="background" tilewidth="32" tileheight="32" tilecount="612" columns="34">
+ <image source="assets/textures/background.png" width="1100" height="600"/>
+</tileset>
