+++ core/src/com/sphy/game/manager/LevelManager.java
@@ -1,0 +1,4 @@
+package com.sphy.game.manager;
+
+public class LevelManager {
+}
