+++ core/src/com/sphy/game/manager/ResourceManager.java
@@ -1,0 +1,4 @@
+package com.sphy.game.manager;
+
+public class ResourceManager {
+}
