+++ mapa1.tiled-session
@@ -1,0 +1,27 @@
+{
+    "Map/SizeTest": {
+        "height": 4300,
+        "width": 2
+    },
+    "activeFile": "",
+    "expandedProjectPaths": [
+        "."
+    ],
+    "fileStates": {
+        "background.tsx": {
+            "scaleInDock": 1,
+            "scaleInEditor": 1
+        }
+    },
+    "last.imagePath": "C:/Users/sanph/XTerminator/assets/textures",
+    "map.fixedSize": true,
+    "map.height": 10,
+    "map.width": 40,
+    "openFiles": [
+    ],
+    "project": "mapa1.tiled-project",
+    "recentFiles": [
+        "background.tsx"
+    ],
+    "tileset.lastUsedFormat": "tsx"
+}
